generators/app/templates/app/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
django-injector==0.1.1
requests[security]==2.24.0
dnspython==2.0.0
boto3==1.14.25
django-health-check==3.12.1
psutil==5.7.2
whitenoise==5.1.0
drf-yasg==1.17.1
sqs-polling==0.1.0